XMonad

Description

Documentation

module XMonad.Main

module XMonad.Core

module XMonad.Config

module XMonad.Layout

module XMonad.ManageHook

module XMonad.Operations

(.|.) :: Bits a => a -> a -> a

Bitwise "or"

class Monad m => MonadState s m | m -> s where

Minimal complete definition

state | get, put

gets :: MonadState s m => (s -> a) -> m a

modify :: MonadState s m => (s -> s) -> m ()

class Monad m => MonadReader r m | m -> r where

Minimal complete definition

(ask | reader), local

asks :: MonadReader r m => (r -> a) -> m a

class Monad m => MonadIO m where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)