XMonad.Config

Description

This module specifies the default configuration values for xmonad.

DO NOT MODIFY THIS FILE! It won't work. You may configure xmonad by providing your own ~/.xmonad/xmonad.hs that overrides specific fields in defaultConfig. For a starting point, you can copy the xmonad.hs found in the man directory, or look at examples on the xmonad wiki.

Documentation

defaultConfig :: XConfig (Choose Tall (Choose (Mirror Tall) Full))

The default set of configuration values itself