XMonad.Layout

Description

The collection of core layouts.

Documentation

data Full a

Simple fullscreen mode. Renders the focused window fullscreen.

Constructors

Full

data Tall a

The builtin tiling mode of xmonad. Supports Shrink, Expand and IncMasterN.

Constructors

Tall

Fields

tallNMaster :: !Int

The default number of windows in the master pane (default: 1)

tallRatioIncrement :: !Rational

Percent of screen to increment by when resizing panes (default: 3/100)

tallRatio :: !Rational

Default proportion of screen occupied by master pane (default: 1/2)

newtype Mirror l a

Mirror a layout, compute its 90 degree rotated form.

Constructors

Mirror (l a)

data Resize

Change the size of the master pane.

Constructors

Shrink

Expand

data IncMasterN

Increase the number of clients in the master pane.

Constructors

IncMasterN !Int

data Choose l r a

A layout that allows users to switch between various layout options.

(|||) :: (LayoutClass l a, LayoutClass r a) => l a -> r a -> Choose l r a

The layout choice combinator

data ChangeLayout

Messages to change the current layout.

Constructors

FirstLayout

NextLayout

mirrorRect :: Rectangle -> Rectangle

Mirror a rectangle.

splitVertically, splitHorizontally :: Int -> Rectangle -> [Rectangle]

splitHorizontallyBy, splitVerticallyBy :: RealFrac r => r -> Rectangle -> (Rectangle, Rectangle)

tile

Arguments

:: Rational
frac, what proportion of the screen to devote to the master area

-> Rectangle
r, the rectangle representing the screen

-> Int
nmaster, the number of windows in the master pane

-> Int
n, the total number of windows to tile

-> [Rectangle]

Compute the positions for windows using the default two-pane tiling algorithm.

The screen is divided into two panes. All clients are then partioned between these two panes. One pane, the master, by convention has the least number of windows in it.