XMonad.Main

Description

xmonad, a minimalist, tiling window manager for X11

Documentation

xmonad :: (LayoutClass l Window, Read (l Window)) => XConfig l -> IO ()

| The entry point into xmonad. Attempts to compile any custom main for xmonad, and if it doesn't find one, just launches the default.

launch :: (LayoutClass l Window, Read (l Window)) => XConfig l -> IO ()

Entry point into xmonad for custom builds.

This function isn't meant to be called by the typical xmonad user because it:

Does not process any command line arguments.

Therefore doesn't know how to restart a running xmonad.

Does not compile your configuration file since it assumes it's actually running from within your compiled configuration.

Unless you know what you are doing, you should probably be using the xmonad function instead.

However, if you are using a custom build environment (such as stack, cabal, make, etc.) you will likely want to call this function instead of xmonad. You probably also want to have a key binding to the restart function that restarts your custom binary with the resume flag set to True.