XMonad.Config

Description

This module specifies the default configuration values for xmonad.

DO NOT MODIFY THIS FILE! It won't work. You may configure xmonad by providing your own ~/.xmonad/xmonad.hs that overrides specific fields in the default config, def. For a starting point, you can copy the xmonad.hs found in the man directory, or look at examples on the xmonad wiki.

Documentation

defaultConfig :: XConfig (Choose Tall (Choose (Mirror Tall) Full))

Deprecated: Use def (from Data.Default, and re-exported by XMonad and XMonad.Config) instead.

The default set of configuration values itself

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.

Orphan instances

a ~ Choose Tall (Choose (Mirror Tall) Full) => Default (XConfig a)

Instance details

Methods

def :: XConfig a