XMonad.Layout

Description

The collection of core layouts.

Documentation

data Full a

Simple fullscreen mode. Renders the focused window fullscreen.

Constructors

Full

data Tall a

The builtin tiling mode of xmonad. Supports Shrink, Expand and IncMasterN.

Constructors

Tall

Fields

tallNMaster :: !Int

The default number of windows in the master pane (default: 1)

tallRatioIncrement :: !Rational

Percent of screen to increment by when resizing panes (default: 3/100)

tallRatio :: !Rational

Default proportion of screen occupied by master pane (default: 1/2)

newtype Mirror l a

Mirror a layout, compute its 90 degree rotated form.

Constructors

Mirror (l a)

data Resize

Change the size of the master pane.

Constructors

Shrink

Expand

data IncMasterN

Increase the number of clients in the master pane.

Constructors

IncMasterN !Int

data Choose l r a

A layout that allows users to switch between various layout options.

Constructors

Choose CLR (l a) (r a)

(|||) :: l a -> r a -> Choose l r a infixr 5

The layout choice combinator

data CLR

Choose the current sub-layout (left or right) in Choose.

Constructors

CL

CR

data ChangeLayout

Messages to change the current layout. Also see JumpToLayout.

Constructors

FirstLayout

NextLayout

newtype JumpToLayout

A message to jump to a particular layout, specified by its description string.

The argument given to a JumpToLayout message should be the description of the layout to be selected. If you use XMonad.Hooks.DynamicLog from xmonad-contrib, this is the name of the layout displayed in your status bar. Alternatively, you can use GHCi to determine the proper name to use. For example:

$ ghci
GHCi, version 6.8.2: http://www.haskell.org/ghc/  :? for help
Loading package base ... linking ... done.
:set prompt "> "    -- don't show loaded module names
> :m +XMonad.Core   -- load the xmonad core
> :m +XMonad.Layout.Grid  -- load whatever module you want to use
> description Grid  -- find out what it's called
"Grid"

As yet another (possibly easier) alternative, you can use the XMonad.Layout.Renamed module (also in xmonad-contrib) to give custom names to your layouts, and use those.

For example, if you want to jump directly to the Full layout you can do

, ((modm .|. controlMask, xK_f), sendMessage $ JumpToLayout "Full")

Constructors

JumpToLayout String

mirrorRect :: Rectangle -> Rectangle

Mirror a rectangle.

splitVertically :: Int -> Rectangle -> [Rectangle]

splitHorizontally :: Int -> Rectangle -> [Rectangle]

splitHorizontallyBy :: RealFrac r => r -> Rectangle -> (Rectangle, Rectangle)

splitVerticallyBy :: RealFrac r => r -> Rectangle -> (Rectangle, Rectangle)

tile

Arguments

:: Rational
frac, what proportion of the screen to devote to the master area

-> Rectangle
r, the rectangle representing the screen

-> Int
nmaster, the number of windows in the master pane

-> Int
n, the total number of windows to tile

-> [Rectangle]

Compute the positions for windows using the default two-pane tiling algorithm.

The screen is divided into two panes. All clients are then partitioned between these two panes. One pane, the master, by convention has the least number of windows in it.