XMonad.ManageHook

Description

An EDSL for ManageHooks

Documentation

liftX :: X a -> Query a

Lift an X action to a Query.

idHook :: Monoid m => m

The identity hook that returns the WindowSet unchanged.

(<+>) :: Monoid m => m -> m -> m

Infix mappend. Compose two ManageHook from right to left.

composeAll :: Monoid m => [m] -> m

Compose the list of ManageHooks.

(-->) :: (Monad m, Monoid a) => m Bool -> m a -> m a infix 0

p --> x. If p returns True, execute the ManageHook.

(-->) :: Monoid m => Query Bool -> Query m -> Query m -- a simpler type

(=?) :: Eq a => Query a -> a -> Query Bool

q =? x. if the result of q equals x, return True.

(<&&>) :: Monad m => m Bool -> m Bool -> m Bool infixr 3

&& lifted to a Monad.

(<||>) :: Monad m => m Bool -> m Bool -> m Bool infixr 3

|| lifted to a Monad.

title :: Query String

Return the window title.

appName :: Query String

Return the application name; i.e., the first string returned by WM_CLASS.

resource :: Query String

Backwards compatible alias for appName.

className :: Query String

Return the resource class; i.e., the second string returned by WM_CLASS.

stringProperty :: String -> Query String

A query that can return an arbitrary X property of type String, identified by name.

getStringProperty :: Display -> Window -> String -> X (Maybe String)

willFloat :: Query Bool

Return whether the window will be a floating window or not

doF :: (s -> s) -> Query (Endo s)

Modify the WindowSet with a pure function.

doFloat :: ManageHook

Move the window to the floating layer.

doIgnore :: ManageHook

Map the window and remove it from the WindowSet.

doShift :: WorkspaceId -> ManageHook

Move the window to a given workspace